# WMS.Angular/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "WMS API v1",
    "version": "v1"
  },
  "paths": {
...
    "/transfers/pallet/locate/number/{palletNumber}/{warehouseId}": {
      "get": {
        "tags": [
          "Transfers"
        ],
        "summary": "Locate a pallet by its pallet number",
        "description": "Finds a pallet by pallet number within a warehouse and returns its current bin location, if checked in.",
        "operationId": "LocatePalletByNumber",
        "parameters": [
          {
            "name": "palletNumber",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "warehouseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PalletLocationDto"
                }
              }
            }
          }
        }
      }
    },
    "/transfers/pallet/locate/qrcode/{hashCode}/{warehouseId}": {
      "get": {
        "tags": [
          "Transfers"
        ],
        "summary": "Locate a pallet by its QR code hash",
        "description": "Finds a pallet by its QR code hash within a warehouse and returns its current bin location, if checked in.",
        "operationId": "LocatePalletByQrCode",
        "parameters": [
          {
            "name": "hashCode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "warehouseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PalletLocationDto"
                }
              }
            }
          }
        }
      }
    },
    "/transfers/pallet": {
      "post": {
        "tags": [
          "Transfers"
        ],
        "summary": "Transfer a whole pallet to a new bin",
        "description": "Moves a checked-in pallet and its contents from its current bin into a new check-in record at the destination bin, logging the transfer.",
        "operationId": "TransferPallet",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferPalletDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransferResultDto"
                }
              }
            }
          }
        }
      }
    },
    "/transfers/item": {
      "post": {
        "tags": [
          "Transfers"
        ],
        "summary": "Transfer a single item to a new bin",
        "description": "Moves a single received product out of its current check-in and into a new check-in at the destination bin, logging the transfer.",
        "operationId": "TransferItem",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferItemDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransferResultDto"
                }
              }
            }
          }
        }
      }
    },
    "/transfers/report": {
      "get": {
        "tags": [
          "Transfers"
        ],
        "summary": "Get a transfer log report",
        "description": "Returns transfer log entries within a date range, optionally scoped to a single warehouse.",
        "operationId": "GetTransferReport",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "warehouseId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TransferLogDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/transfers/item/bin/{binId}": {
      "get": {
        "tags": [
          "Transfers"
        ],
        "summary": "List individually checked-in items in a bin",
        "description": "Returns item-level (non-pallet) received products currently checked in to the given bin.",
        "operationId": "GetItemsInBin",
        "parameters": [
          {
            "name": "binId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ItemLocationSummaryDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/pallet": {
      "get": {
        "tags": [
          "WMS.Api"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PalletSummaryDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "WMS.Api"
        ],
        "summary": "Create a new pallet",
        "description": "Creates a new pallet record in the system and returns its details.",
        "operationId": "CreatePallet",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePalletDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PalletSummaryDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreatePalletDto": {
        "type": "object",
        "properties": {
          "warehouseId": {
            "type": "integer",
            "format": "int32"
          },
          "palletNumber": {
            "type": "integer",
            "format": "int32"
          },
          "palletHashCode": {
            "type": "integer",
            "format": "int32"
          },
          "palletDimension": {
            "type": "string",
            "nullable": true
          },
          "capacityInKG": {
            "type": "number",
            "format": "double"
          },
          "capacityInCBM": {
            "type": "number",
            "format": "double"
          },
          "dateAdded": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "nullable": true
          }
        },
        "additionalProperties": {}
      },
...
      "ItemLocationSummaryDto": {
        "type": "object",
        "properties": {
          "receivedProductId": {
            "type": "integer",
            "format": "int32"
          },
          "checkInId": {
            "type": "integer",
            "format": "int32"
          },
          "productName": {
            "type": "string",
            "nullable": true
          },
          "typeOfPackage": {
            "type": "string",
            "nullable": true
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "receivingSeries": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "PalletLocationDto": {
        "type": "object",
        "properties": {
          "palletId": {
            "type": "integer",
            "format": "int32"
          },
          "warehouseId": {
            "type": "integer",
            "format": "int32"
          },
          "palletNumber": {
            "type": "integer",
            "format": "int32"
          },
          "palletHashCode": {
            "type": "integer",
            "format": "int32"
          },
          "checkInId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "currentBinId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "currentBinLocation": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PalletSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "warehouse": {
            "type": "string",
            "nullable": true
          },
          "palletNumber": {
            "type": "integer",
            "format": "int32"
          },
          "palletHashCode": {
            "type": "integer",
            "format": "int32"
          },
          "palletDimension": {
            "type": "string",
            "nullable": true
          },
          "capacityInKG": {
            "type": "number",
            "format": "double"
          },
          "capacityInCBM": {
            "type": "number",
            "format": "double"
          },
          "dateAdded": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "TransferItemDto": {
        "type": "object",
        "properties": {
          "receivedProductId": {
            "type": "integer",
            "format": "int32"
          },
          "fromCheckInId": {
            "type": "integer",
            "format": "int32"
          },
          "toBinId": {
            "type": "integer",
            "format": "int32"
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TransferLogDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "transferType": {
            "type": "string",
            "nullable": true
          },
          "itemDescription": {
            "type": "string",
            "nullable": true
          },
          "fromBinLocation": {
            "type": "string",
            "nullable": true
          },
          "toBinLocation": {
            "type": "string",
            "nullable": true
          },
          "transferDate": {
            "type": "string",
            "format": "date-time"
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TransferPalletDto": {
        "type": "object",
        "properties": {
          "palletId": {
            "type": "integer",
            "format": "int32"
          },
          "toBinId": {
            "type": "integer",
            "format": "int32"
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TransferResultDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "CookieAuth": {
        "type": "apiKey",
        "description": "ASP.NET Core Identity Application Cookie",
        "name": ".Skybest.WMS.Auth",
        "in": "cookie"
      }
    }
  },
  "security": [
    {
      "CookieAuth": []
    }
  ]
}